Base Layer explanation and adding user's intention
I noticed that the small (but important) notion of "keeping objects on the base layer" disappeared. So, I added it back.

I also changed some words to match user's search and reading intention: *following* an object vs *moving* an object, *camera* vs *layer* (I changed the word "layers" for "objects" because even if it's technically correct, the intention of the reader is on an object, not on a layer). 

I "deprioritised" the use of other extensions like "copy camera settings", but I still included it under the "follow smoothly" section.